consolidate and align url stuff
