+++ frontend/src/styles/HetComponents/HetDesktopToolbar.tsx
@@ -1,14 +1,15 @@
 import { Menu, MenuItem, Toolbar } from '@mui/material'
 import { useState } from 'react'
 import AppBarLogo from '../../assets/AppbarLogo.png'
+import { NAVIGATION_STRUCTURE } from '../../pages/navigationData'
 import { EXPLORE_DATA_PAGE_LINK } from '../../utils/internalRoutes'
-import { NAVIGATION_STRUCTURE } from '../../utils/urlutils'
+import { isExternalLink } from '../../utils/urlutils'
 import HetCTASmall from './HetCTASmall'
 import HetLaunchLink from './HetLaunchLink'
 import HetNavButton from './HetNavButton'
 import HetNavLink from './HetNavLink'
 
-export default function HetAppToolbar() {
+export default function HetDesktopToolbar() {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
   const [activeMenu, setActiveMenu] = useState<string | null>(null)
 
@@ -42,11 +43,7 @@ export default function HetAppToolbar() {
               classes={{ paper: 'bg-white' }}
             >
               {Object.entries(value.pages).map(([subKey, subValue]) => {
-                const isExternal =
-                  typeof subValue === 'object'
-                    ? subValue.isExternal
-                    : subKey.startsWith('https://') ||
-                      subKey.startsWith('http://')
+                const external = isExternalLink(subKey)
                 const label =
                   typeof subValue === 'string' ? subValue : subValue.label
 
@@ -55,14 +52,14 @@ export default function HetAppToolbar() {
                     <div className='flex items-center gap-2'>
                       <HetNavLink
                         href={subKey}
-                        {...(isExternal && {
+                        {...(external && {
                           target: '_blank',
                           rel: 'noopener noreferrer',
                         })}
                       >
                         {label}
                       </HetNavLink>
-                      {isExternal && (
+                      {external && (
                         <HetLaunchLink
                           svgClassName='flex my-auto text-text'
                           href={subKey}
@@ -78,24 +75,21 @@ export default function HetAppToolbar() {
       }
 
       if ('link' in value) {
-        const isExternal =
-          value.isExternal ||
-          value.link.startsWith('https://') ||
-          value.link.startsWith('http://')
+        const external = isExternalLink(value.link)
 
         return (
           <div key={key} className='flex items-center gap-1'>
             <HetNavLink
               href={value.link}
               className='mx-2 my-0 w-auto px-2 font-medium font-sans-title text-navlink-color text-small'
-              {...(isExternal && {
+              {...(external && {
                 target: '_blank',
                 rel: 'noopener noreferrer',
               })}
             >
               {value.label}
             </HetNavLink>
-            {isExternal && <HetLaunchLink href={value.link} />}
+            {external && <HetLaunchLink href={value.link} />}
           </div>
         )
       }
